requirements/base.txt
Flask>=3.0.0,<4.0.0
Flask-Caching>=2.1.0,<3.0.0
Flask-Limiter>=3.5.0,<4.0.0
Flask-RESTX>=1.3.0,<2.0.0
Flask-SQLAlchemy>=3.1.1,<4.0.0
Flask-Migrate>=4.0.5,<5.0.0
gunicorn>=21.2.0,<22.0.0
python-dotenv>=1.1.0,<2.0.0
pydantic>=2.11.3,<3.0.0
SQLAlchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0
redis>=5.2.1,<6.0.0
prometheus-flask-exporter>=0.23.0,<1.0.0
pytest>=8.3.5,<9.0.0
Werkzeug>=3.0.0,<4.0.0
rich>=12.0.0,<13.0.0
loguru>=0.7.2,<1.0.0
rq>=1.15.1,<2.0.0 
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
huggingface_hub>=0.20.0,<1.0.0
pyannote.audio>=3.1.1,<4.0.0
Flask-Cors>=4.0.0,<5.0.0 
faster_whisper>=0.1.0,<1.0.0
